vim: filetype detection for vimdiff as svn diff
If vimdiff is used as the diff program for "svn diff", it will be run
with a file in .svn/text-base as the left file.  That file will have
.svn-base appended to the actual file name, preventing file type
detection based on the extension.  Strip of the extra extension and try
again.

Not using svn much anymore, but this is a pretty nice hack.  Could also
be useful as an example of other ways to redo filetype detection.