vim - Set path to search for tags files
First search vim's CWD, then the directory of the current file and its
ancestors.  That reverses the standard order, which doesn't include
ancestors either.

I often work in subdirectories of a project, and want to have a single
tags file for the entire thing.  The standard setting doesn't work in
this case.