vim - Avoid double setting of listchars
Had been setting it for both latin1 and utf8 encodings, switch to just
stating that the vimrc file is encoded as utf8 and let vim do the
conversion the rare times that it's necessary.

Previously the latin1 version prevented the file being seen as utf8, and
the utf8 versions then showed as garbage when editing.