zsh - Don't set $EMAIL
Having this set will cause git to use that value for my committer ID,
preventing it from complaining if I haven't set it for a repo where I'm
doing a commit.  Since I want to use different addresses based on the
repo, having a single setting in the environment won't work.