zsh - Stop setting TERM=gnome-256color
Unfortunately vim seems to only enable mouse support for a small number
of $TERM settings, not including gnome-256color.

I don't recall what problems I was having with using TERM=xterm.  I
think that it was issues with the tmux status line, and I'm not seeing
those currently.  So go back to using that.  Even if there are minor
problems the problem of vim not letting me use the mouse is arguably
worse, although it is strange that it apparently took me over 6 months
to notice this.