xsessionrc - added
Add xsessionrc script to startup gpg-agent to server for both gpg and
ssh.  It does a better job of being an agent for ssh than ssh-agent
does: once it's told about an ssh key it will remember that even across
restarts, so it rather than ssh itself will be the one that asks for the
pass phrase to unlock a key that hasn't been used in a session.