vim - Disable perl_module plugin
Came across a filename that contained "::", this plugin prevented me
from editing it.  The autocommand pattern syntax is too limited to
really make this plausible.

Could perhaps have it call a function to do better checking of the
filename, and check if a file with the requested name actually exists.