vim - Use try/catch around pathogen load
Use try/catch to check if there are errors loading pathogen instead of
just checking if the expected file is readable.  This will be more
portable in case I'd happen to want to use vim on Windows where the file
wouldn't be inside ~/.vim, and it will also have the opportunity to
handle other errors.