vim - Turn on filetype outside pathogen block
Previously the pathogen block was turning filetype support off, then on
again because that used to be necessary.  But, currently
pathogen#infect() does that internally if the support was previously on.

But, the support for filetype plugins and indenting still needs to be
enabled.  Do that after pathogen has loaded, so that if run on a system
where filetype support isn't already turned on it isn't done until after
pathogen, removing the need for that to bounce that flag and perhaps
cutting down on the load time slightly.