vim - Trim more from statusline format info
When file format and encoding match their usual values, remove the
leading "," as well to save a couple more characters in the common case.
Switch to using ";" as the separator in front of the encoding, as that's
also the separator that would be used in a MIME Content-Type header
before the encoding parameter.