zsh - glob absolute and relative paths for scp
Modify scp function so that an argument is not considered to be a remote
path reference if the colon comes after a slash since hostnames cannot
contain slashes.  This mimics behaviour of scp itself, and allows
prefixing a path with "./" to refer to a filename which happens to
contain a colon.