zsh - Fix vi mode display after history editing
When in zle-line-finish hook call zle-mode function with "finish"
parameter.  Modify that function so that when it is called in that way
it will avoid saying that the line editor is in vi command mode.  This
avoids displaying the incorrect mode after exiting the editor in that
mode such as when selecting a line from the history.