mutt - No Mail-Followup-To to git list
Attempt to avoid adding a Mail-Followup-To header when sending messages
to the git mailing list, since some people there will complain about it
and it generally won't be honored anyway.

There isn't really a way to guarantee that the header isn't used.  My
first attempt was to use a send-hook to unset the followup_to variable
when sending to the list address, but that would often fail if I
postponed a message then recalled it after sending a message elsewhere
which isn't unlikely.

My current solution is to use a folder-hook to unset the variable when
in my folder for the mailing list.  This is much less likely to fail
given how I typically send messages to mailing lists.