vim - Turn off joinspaces for markdown files
The joinspaces option causes vim to use two spaces at the end of what it
thinks is a sentence.  When that is also the end of a line that causes
stackoverflow to treat it as the end of a paragraph which can result in
ugly line breaks.  Since this type of whitespace isn't important in
markdown files anyway turn the option off to get just a single space.