ssh - share qqx.org config with proxy
Modify settings for *.qqx.org and "proxy" so that settings from the
former will be used with the latter as well by default.  This eliminates
the need to separately specify the port to use for the ssh server,
especially nice when I take my laptop somewhere with restricted wifi and
I need to use port 443 to connect to pug.