vim - Remove command-t plugin
The repo URL that I'd had for the command-T plugin is currently not
working causing problems with setting up my dotfiles on new boxes.  I
found a mirror of it, but updating to the latest version from that still
resulted in a copy that would segfault when I try to use the plugin.
Since it's been broken for ages, just remove it.