i3 - Move bars to top
Move i3 status bars to the top of the screen.  On my new laptop the
pointer moves to the bottom of the screen when the touchpad is first
used after using the touch screen.  In this situation having the status
bar at the bottom of the screen causes an attempt to scroll to switch
workspaces even though the cursor had been elsewhere.