vim - Eliminate spaces in airline cursor position
Modify setting of section of airline used to display the current cursor
position to eliminate spaces.  Don't set a width for the number of lines
in the file (the number of digits needed for that changes rarely enough
that it shouldn't be an issue), and use left alignment for the column
number.