zsh - Add burt_summary_db function
When I run burt_summary_db directly I'll always want the
BURT_SUMMARY_DB_GET_SEP environment variable set, but I don't want to
set that in the environment generally in case it would cause problems
for other commands that invoke that.  Use a shell function wrapper to
set it only when I run it from the shell.