vim - Fixed version of editing perl module by name
Had previously commented BufReadCmd autocommand which would attempt to
find perl module for any file where the name contained "::", since that
prevented editing files which actually matched that pattern.

Now have an improved version which first checks to see if such a file
exists, and just reads it if found.