zsh - Be careful about using url-quote-magic
Only setup bindings for url-quote-magic if marking the function for
autoload was successful.  If the function can't be autoloaded that
results in a completely unusable shell, since nothing can be typed into
it.  This was happening on box which was mistakenly using an old version
of zsh which doesn't support one of the options used for the autoload
command.