screen - go back to using my usual prefix
I now need to use screen rather than tmux on some machines where the
latter is broken, so switch back to using ^O as the prefix so I have
less need to recall if I'm using screen or tmux.  This will likely be
more common than using screen inside of tmux which was why I'd
previously commented this out.