vim - Move SameSyntaxMotion plugin into bundle dir
Use pathogen to manage SameSyntaxMotion plugin even though it isn't
being managed as a git submodule since it doesn't seem to be available
from a public git repo.  This will automatically prevent it from loading
if pathogen isn't setup yet or is disabled for some reason, and that's
an improvement since it depends on a plugin which was already managed as
a submodule and with pathogen.