zsh - Don't unset TMPDIR for tmux
Stop unsetting $TMPDIR in tmux alias.  Using a TMPDIR that is different
than the default causes zsh completion of session names to fail, since
that uses `command tmux` to get the list, avoiding use of the alias.
I'd originally started doing that to allow me to use tmux as a login
shell, but I've since decided against doing that.