zsh - update time in prompt
Call `zle reset-prompt` every second to update the time.  My main reason
for including the time down to the second in the prompt is so that I can
get a good idea of how long commands take without needing to do anything
up beforehand, but that is doesn't work if commands are started with a
stale timestamp in the prompt.