bash - changes to setting of PS1
Only modify PS1 if it's already set.  That being set is a common way of
determining if a shell is interactive, so it shouldn't be set
unconditionally.

Reset color after the first line of the prompt, not doing so sometimes
causes the prompt color to be used for non-prompt content.