mutt - macro to toggle auxiliary sort method
Create a macro bound to 'S' in the index to toggle between sorting
threads based on individual sent dates or the last received date.  In
general it's nice to have new messages near each other as happens with
the latter method, but it can occasionally make it harder to follow
along with a conversation when it causes me to read references to
messages that I haven't yet gotten to.  Using the sent date can also
help to display series of patches in the correct order.