vim - Always show tab line in GUI
When the tab line is added, the GTK UI assumes that it can resize the
window and so it doesn't adjust the size of the content area if that
fails (like when being run in a tiling window manager).  As a result
this pushes the command line and most of the status line off of the
screeen.  Work around this by always showing the tab line in the GUI.