cropmail - Filter messages addressed to git list
Since the policy of the git mailing list is to aggressively copy anybody
involved in a thread, it can lead to tons of messages going to my inbox
anytime I participate.  Filter messages addressed to the list as if they
came through the list; I've already enabled duplicate checking for it.