zsh - Add space to end of first line of prompt
The character that I include in my prompt when there are unstaged
changes to the git repo is too wide to fit in a single cell in iTerm,
and the right portion doesn't display when the background gets reset
after the prompt.  Add a space after this so that the background will
extend farther and have the entire character displayed.