git - Fix inclusion of local global config
Putting the real copy of my global git config file caused it to no
longer find the local (non-version-controlled) file to include. Update
the include directive to fix this. It will now look for both the old
name in the home directory and a file named simply `local` in the XDG
directory.