vim - unimpaired-style mappings for ale
Use `[s` and `]s` (mnemonic: syntax), along with capitalized versions to
move between errors found by ale plugin.  The unimpaired mappings to use
the quickfix or location list don't work well for this, since ale keeps
resetting the underlying list which causes those mappings to go back to
the start of the file.