vim - Disable guicursor for nvim
When guicursor option is set in a neovim session within konsole it
causes the font size to be reset to the terminal default, including
immediately after the font size is changed. This is because the method
of changing the cursor there involves temporarily switching to a
temporary profile.

Avoid this by clearing this option for nvim. Will need to revisit this
if I start using a neovim GUI.