mutt - Fix reply_regexp
Way back in 4abea3bc from 23 Feb 2010, I modified $reply_regexp to
attempt to deal with subjects in some mailing lists getting prefixes
like `[tag:###]`. But somehow failed to notice that the change was much
broader than that. Change it again to require both the tag and number be
present to tighten it up.

Finally realized that because some messages with a subject of
`[blacklisted]` were showing up as having no subject. But I know that a
number of other messages had WAY too much removed.