zsh - Avoid $TERMINFO pointing to single-entry database
When `kitty` starts on Mac OS it sets `$TERMINFO` to point to it's own
terminfo entry, if the tmux server is started with that set it won't be
able to work inside of terminals that use any other value for `$TERM`.
Detect that situation, and reset the variable to point to home directory
instead. The database there can get a copy of the entry for kitty with:

    tic =(infocmp xterm-kitty)