tmux - More efficient conditional for vim tmux navigator
At least on Mac (as of some recent upgrade) using a shell command to
check if the current pane is running vim is horribly (pretty much
unusably) slow.  Newer versions of tmux support pattern matching in
format strings and using the results for conditions, so use that when
tmux is at least 3.0.