dot_setup - Ignore trailing slashes
Remove trailing slash before creating parent directories for
destination. When a directory name is specified for setup, the desire
would be to have a symlink to that directory. But if a trailing slash
were included, we'd first create the destination as en empty directory;
after that attempting to create the symlink would fail due to the
destination already existing as a directory.