vim - Add mapping for inserting semicolon
If semicolon is entered as the first non-whitespace character of the
line, move the cursor left after entering it. This allows the line
terminator for C, perl, Java, and (some) JavaScript to be entered before
the command that it terminates so that there will be less time when the
file is syntactically invalid.