zsh - Add `@` function
This goes along with the previous commit. While I now view the `Equals`
option as a bit too wide reaching for my tastes I'd still like something
fairly short that I can use to replace it. This will allow me to use
`$(@ program)` (or the equivalent with backticks) where I'd have
previously used `=program`.  This is still somewhat longer, but not by a
horrible amount.