resolve - Try to resolve symlinks one step at a time
If trying to resolve something that goes through Debian's alternatives
system, I want to see that; but unfortunately perl's abs_path function
will fully resolve to the final location hiding the trip through
/etc/alternatives. When the `realpath` command is present use that along
with `readlink` to resolve symlinks one step at a time so that this is
shown. If `realpath` isn't available (such as a default Mac OS
installation without coreutils installed) fallback to using perl.