sh - Convert @ from alias to a script
Thinking on this further, I really only want this to look for command in
`$PATH`, not aliases or functions; so there's no reason that this should
be an alias.

This *could* be just a simple wrapper around `which`, but I wanted
something a bit more verbose. Also, the recent Debian moves around
deprecating the usual `which` there make me not want to use that.