zsh - Ignore lam(1) for completion
I've never wanted to use this, didn't even know about it, and it has
repeatedly gotten in the way of completing my own lambda-urls command.
This is also short enough that completion doesn't really even help for
it.

While here, do a bit of restructuring of the list of patterns to ignore
to allow comments on specific items.