direnv - Support additional path dir in load_brew_prefix
Recently Homebrew's python packages were modified so that the `bin`
directory only includes executables that include the minor version (e.g.
`python3.7`), but the intent with load_brew_prefix was to be able to
specify a python version there, and then have `python3` pick up the
minor version from that. Those packages *do* still have the more
loosely-versioned executables under `libexec/bin`, so modify the helper
to be able to include that in the path as well.

So, now my `.envrc` files for python repos will look more like:

    load_brew_prefix 'python@3.7' libexec/bin