tmux - Add prefix bindings for nav keys
For a long time I've had the ^H, ^J, ^K, and ^L bound to navigate
between tmux panes and vim windows, and that works nicely. But,
occasionally I may want to run a program inside of tmux that uses one of
those for something important; so add (^O) prefixed bindings that will
pass the keystroke through.