st - Use exit rather than return
This was originally written to be used as a shell function, so I needed
to use `return` rather than `exit` to avoid ending an interactive shell.
I later switched to using it as a separate script for easier updates,
and since there isn't really a need for the slight optimization of use
as a function. When I did that I neglected to convert to using `exit`,
and I recently started noticing warnings from bash about that.