zsh - New filename for syntax highlighting plugin
When I updated this plugin in 9480c55 (zsh - Update syntax highlighting
plugin, 2023-07-27) I failed to notice that the actual reason errors
went away was that I was no longer loading the plugin due to a changed
file name. Adjust to the new name of the plugin file
(`F-Sy-H.plugin.zsh`).