zsh - Fix binding tab to complete-word
In be46d10 (zsh - Bind tab to complete-word, 2023-07-27) I swapped the
arguments to `bindmodes` so that I wasn't actually changing the binding
for the tab key. This was the actual reason that I started getting the
errors which I thought I was fixing with 9480c55 (zsh - Update syntax
highlighting plugin, 2023-07-27).