vim - Hack to help indenting in pug files
For some reason the plugin for pug files (pulled in through polyglot)
decided to set it up so when the `<Return>` key is pressed the indent
for the line being left would be recalculated. At least for me this very
often breaks things, so add a hack to disable it.

I don't think there's really a way to determine when the indent of a
line should be different than for the previous in pug files, except in a
few cases (e.g. `if`, `each`).