shell - Fix update_env_from_tmux() to work in bash
Rather than piping to a `while read` statement, use process substitution
to feed the data. The previous method is fine in zsh, but in bash it
causes the looping to be done in a subshell, which makes the effects
worthless. The new method appears to work correctly in both bash and
zsh.