resolve - Try to use GNU realpath(1) on Mac OS
As noted in 8345f548 (resolve - Fix for Mac OS, 2024-02-15), Mac OS now
has a version of realpath(1) that doesn't support the `-s` option.
Homebrew has a `coreutil` package that includes GNU version, but the
usual $PATH only has it with a `g` prefix; try to use that if the normal
name won't work.