vim - Don't require python support for nvim-cmp
Back when I was using deoplete, that required nvim with python3 support
and so I was blacklisting the plugin when either of those wasn't
available. Then when I started using `nvim-cmp` I just plugged that in
the same place even though it doesn't require python support. I no
longer use deoplete anywhere, so remove the blacklisting of that and the
check for python support.