vim - Set python host earlier
Move up the addition from e8b3b2ed (vim - Use custom venv for python
host when available, 2024-02-28), since that needs to be done before
plugins are loaded. That should probably be done for the similar node
setup as well, but currently that happening later doesn't seem to be an
issue.