update-certificates - New script
Add script to generate/update a certificates bundle.

The Python `requests` library used by httpie(1) by default verifies
against the CA store provided by the `certifi` package rather than any
system store. That makes it not work with locally trusted certificates.
It supports an environment variable to use a different CA bundle, but
that then needs the usual certificates. Create a script to create a
bundle file with the system CA certificates as well as personally
trusted CAs. For now this only gets system CAs on Mac OS, but it should
be easy to adapt for at least Debian-based systems.