vim - Fix tmux_navigator with mail files
I have tmux configured to look for "Vim@" in the window title to
determine if navigation keys should be passed to vim for the
tmux_navigator plugin to handle. I also had an ftplugin for the `mail`
file type to change the window title to the email subject which broke
that. Change the ftplugin to instead set a buffer variable with the
subject, then have that used instead of the file path for the
titlestring. Besides fixing navigating away from email buffers, this
also brings back consistency.